requirements.txt
# FastAPI and ASGI server
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis[hiredis]==5.0.1

# Storage (MinIO for local, Azure for production)
minio==7.2.3
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.0

# Payments
razorpay==1.4.1

# Background tasks
celery[redis]==5.3.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14

# Misc
setuptools>=69.0.0firebase-admin==7.1.0
